requirements.txt
Pillow==9.5.0
rembg==2.0.35